dev-requirements.txt
-r requirements.txt
blessings==1.7
bpython==0.18
django-extensions==2.2.9
django-plainpasswordhasher==0.3

